Refactor Visitor
